+++ video/img_format.h
@@ -252,6 +252,8 @@ enum mp_imgfmt {
     IMGFMT_VDPAU_FIRST = IMGFMT_VDPAU,
     IMGFMT_VDPAU_LAST  = IMGFMT_VDPAU_MPEG4,
 
+    IMGFMT_VDA,
+
     IMGFMT_VAAPI,
     IMGFMT_VAAPI_MPEG2_IDCT,
     IMGFMT_VAAPI_MPEG2_MOCO,
@@ -335,10 +337,13 @@ static inline bool IMGFMT_IS_RGB(unsigned int fmt)
 #define IMGFMT_IS_VDPAU(fmt) \
     (((fmt) >= IMGFMT_VDPAU_FIRST) && ((fmt) <= IMGFMT_VDPAU_LAST))
 
+#define IMGFMT_IS_VDA(fmt) (((fmt) == IMGFMT_VDA))
+
 #define IMGFMT_IS_VAAPI(fmt) \
     (((fmt) >= IMGFMT_VAAPI_FIRST) && ((fmt) <= IMGFMT_VAAPI_LAST))
 
-#define IMGFMT_IS_HWACCEL(fmt) (IMGFMT_IS_VDPAU(fmt) || IMGFMT_IS_VAAPI(fmt))
+#define IMGFMT_IS_HWACCEL(fmt) \
+    (IMGFMT_IS_VDPAU(fmt) || IMGFMT_IS_VAAPI(fmt) || IMGFMT_IS_VDA(fmt))
 
 
 struct mp_imgfmt_entry {